Fixing type mismatch exceptions

+++ Quay-Gnomes/src/main/java/Controllers/Basket.java
@@ -30,7 +30,7 @@ public final class Basket implements Serializable {
 
 
 
-	private double totalPriceOfBasket;
+	private float totalPriceOfBasket;
 	ItemInBasket item;
